Refactor authschema
