+++ src/api/appApiClient.js
@@ -1,6 +1,6 @@
 import axios from 'axios';
 import baseURL from '../config';
 
-export default axios.create({
+export default appApiClient = axios.create({
   baseURL,
 });
